--- src/lib/balance/index.js
@@ -1,1 +1,0 @@
-export { default as balanceStore } from "./balanceStore";
