+++ src/lib/components/__tests__/Balance.spec.js
