+++ src/lib/components/index.js
@@ -1,0 +1,1 @@
+export { default as Balance } from "./Balance/Balance.svelte";
