--- src/lib/contracts/index.js
@@ -1,1 +1,0 @@
-export { default as operationsStore } from "./operationsStore";
