+++ src/lib/dusk/components/Mnemonic/enDictionary.js
