--- src/lib/settings/index.js
@@ -1,1 +1,0 @@
-export { default as settingsStore } from "./settingsStore";
