+++ src/lib/stores/index.js
@@ -1,0 +1,4 @@
+export { default as balanceStore } from "./balanceStore";
+export { default as onboardingWizardStore } from "./onboardingWizardStore";
+export { default as operationsStore } from "./operationsStore";
+export { default as settingsStore } from "./settingsStore";
