+++ src/lib/stores/settingsStore.js
